1、微信自动发到程序员自己的微信用户 2、加入判断源文件编码，避免读写文件出错
